Reorganized controllers
