+++ backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/controller/auth/AuthenticationRequest.kt
@@ -1,4 +1,4 @@
-package me.camwalford.backendapiservice.dto
+package me.camwalford.backendapiservice.controller.auth
 
 data class AuthenticationRequest(
     val username: String,
