+++ backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/controller/sentiment/SentimentController.kt
@@ -1,9 +1,5 @@
-package me.camwalford.backendapiservice.controller
+package me.camwalford.backendapiservice.controller.sentiment
 
-import kotlinx.coroutines.withContext
-import me.camwalford.backendapiservice.context.SecurityContextElement
-import me.camwalford.backendapiservice.dto.SentimentRequest
-import me.camwalford.backendapiservice.dto.SentimentResponse
 import me.camwalford.backendapiservice.service.SentimentAnalysisService
 import me.camwalford.backendapiservice.service.UserService
 import me.camwalford.backendapiservice.service.CompanyNewsService
@@ -11,7 +7,6 @@ import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
-import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.web.bind.annotation.*
 import org.springframework.web.server.ResponseStatusException
