+++ backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/controller/sentiment/SentimentResponse.kt
@@ -1,0 +1,6 @@
+package me.camwalford.backendapiservice.controller.sentiment
+
+data class SentimentResponse(
+    val results: List<SentimentResult>,
+    val creditsRemaining: Int
+)
