+++ backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/controller/sentiment/SentimentResult.kt
@@ -1,4 +1,4 @@
-package me.camwalford.backendapiservice.dto
+package me.camwalford.backendapiservice.controller.sentiment
 
 import com.fasterxml.jackson.annotation.JsonProperty
