--- backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/dto/SentimentResponse.kt
@@ -1,8 +1,0 @@
-package me.camwalford.backendapiservice.dto
-
-import me.camwalford.backendapiservice.model.SentimentData
-
-data class SentimentResponse(
-    val results: List<SentimentResult>,
-    val creditsRemaining: Int
-)
