+++ backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/service/SentimentAnalysisService.kt
@@ -1,7 +1,7 @@
 package me.camwalford.backendapiservice.service
 
 import me.camwalford.backendapiservice.dto.CompanyNews
-import me.camwalford.backendapiservice.dto.SentimentResult
+import me.camwalford.backendapiservice.controller.sentiment.SentimentResult
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
